[Fix, Feat] 댓글, 답글 CRUD api 연결, 기능 구현 (#130)
* [Fix] href

* [Chore] 컴포넌트 파일 위치 변경

* [Feat] Suspense를 이용해 dynamic import 시 대체 UI 표시

* [Fix, Feat] 공통 컴포넌트 업데이트

* [Chore] 이미지 수정

* [Feat] 전체 화면 로딩 컴포넌트 추가

* [Chore] 프로필 스키마 타입 추가

* [Refactor, Chore] 아이콘 버튼들 분리, 테스트중

* [Chore, Style] 스타일 및 이름 수정

* [Chore] handleAlert 파일 위치 변경으로 경로 수정

* [Fix] 세션 프로바이더 삭제

* [Chore]

* [Feat] crud 테스트중

* [Chore] defaultValue blocked

* [Chore] 프로필 미리보기 삭제

* [Style] loading spinner 사용을 위한 공통 로딩 컨테이너 분리

* [Chore] rm unused component

* [Chore] 파일 위치 변경

* [Chore] 중복 session 검사 제거 => ProfileForms에서 session 검사

* [Chore] rm unused size

* [Chore] p 태그 아래 중첩 태그 오류 방지를 위해 div 태그로 변경

* [Refactor, Feat]

* [Chore] rm unused code block

* [Fix] document is not definded 오류 방지로 LoadingContainer 조건 추가

* [Feat] 계정 삭제 확인 모달 추가

* [Refactor, Fix] 관심 카테고리 선택 단계 분리

* [Refactor] toast 라이브러리 사용시 handleAlert로 대체

* [Chore]

* [Feat] 회원 데이터 삭제 액션 추가

* [Feat, Fix] JSDoc 추가, 프로필 정보 업데이트 오류 수정

* [Refactor, Style] props 전달 삭제, 이벤트 함수 내부에서 정의 후 실행, 스타일 수정

* [Refactor, Feat] 회원 가입 후 관심카테고리 선택 페이지 추가

* [Feat, Fix] 프로필 정보 오류 수정

* [Feat] 뒤로 돌아가기 버튼 추가

* [Chore] 카테고리 그룹 데이터 분리

* [Refactor, Feat] 메인 검색창 컴포넌트 분리, 스터디 검색 결과로 이동

* [Chore] 탭 버튼 클릭 핸들러 다시 클릭시 값 없음 테스트

* [Feat, Fix] 필터 재선택 혹은 전체 선택 시 해당 필터링 취소하도록 수정

* [Feat] 후에 프로필의 관심 카테고리, 가장 최근에 만들어진 스터디들 데이터 가져와서 보여주기

* [Style] GridField 그리드 무너짐 해결 스타일 추가

* [Feat] 관련 스터디 링크 검색, 선택 테스트중

* [Fix, Refactor] 로그인한 사용자 프로필 이미지 연결

* [Feat] 더미 액션 추가

* [Feat] 미션 보드 삭제, 참여 신청한 스터디 목록 영역 추가

* [Feat] 관련 스터디 링크 select로 검색, 선택하여 링크 추가

* [Chore, Refactor] 유틸 함수 이동

* [Feat] 유더 데이터 가져오는 액션 추가

* [Fix] 에러 메시지 수정

* [Chore] 액션, formData 이름 수정

* [Fix] 글 상세 데이터 적용

* [Chore, Style]

* [Fix] 프로필 업데이트 기능 오류 수정

* [Fix] rm type error, change temp error text

* [Refactor] 글로벌 유틸 함수 위치 변경

* [Refactor, Feat] es-hangul 라이브러리를 이용하여 초성 검색 기능 추가

* [Refactor, Style, Fix] 버튼 컴포넌트 리팩토링, 호버 인터랙션 스타일 미적용 문제 해결

* [Fix] 공통 컴포넌트 내부에서 세션 외 데이터를 가져오지 않도록 수정

* [Fix] 댓글 입력 창에서 현재 로그인한(작성중인) 사용자 프로필 표시

* [Chore] 머지 되지 않은 부분 다시 덮어쓰기

* [Fix] 커뮤니티 글 작성 시 관련 스터디 링크 값 전달 수정 중, 유저 데이터 타입 통일중

* [Fix] 커뮤니티 글 작성, 삭제 수정 - 관련 스터디 데이터 입출력 확인 완료

* [Refacor Chore] accept 값 수정, type 위치 변경

* [Chore]

* [Feat] 조회수 increase 액션 수정

* [Feat] 커뮤니티 글 정렬 로직 수정

* [Style] 클래스 선택자 수정

* [Fix] 사용자 데이터 타입 통일, 수정, Profile 스키마 deprecated로 액션 수정 `\lib\schema.ts`

* [Chore] import missing Link

* [Chore, Fix]

* [Feat] 작성자 본인일때 커뮤니티 글 업데이트

* [Refactor] 썸네일 컴포넌트 통일, props 수정

* [Refactor] 로고 svg 컴포넌트 대체

* [Feat] 찜한 스터디, 지원한 스터디, 좋아요한 글 목록 저장을 위한 스키마 추가

* [Chore] 설명 추가

* [Feat] 로그인한 사용자 메인 스터디 현황 보드 데이터 연결, 찜한 스터디 목록 가져오기 액션 추가

* [Feat] 대댓글 길이까지 계산하는 유틸 함수 추가, 사용

* [Style, Fix] 오타 수정

* [Chore] custom fetch 함수 설정

* [Fix] radio name, value 속성 수정

* [Fix] init 속성 optional 변경

* [Chore]

* [Chore] 로그인 성공 시 메인 페이지가 아닌 이전 페이지로 이동

* [Feat, Fix] 커뮤니티 글 API CRUD 수정, 기능 테스트 완료

* [Fix] userId가 없을 때 getAllStudies()

* [Chore]

* [Fix] 다른 페이지에서 로그인 후 redirect 방법 다시 생각해보기

* [Fix] createdAt 갱신 대신 timestamp 사용으로 createdAt, updatedAt 관리

* [Fix] svg dot path 생략되는 문제 수정

* [Refactor] post 데이터 작성자 populate select 범위 수정

* [Chore] 사용하지 않는 import 삭제

* [Feat] 댓글, 답글 CRUD 기능 구현

* [Chore] 비공개 라우트 제거, 준비되지 않은 페이지 안내

* [Fix] reply가 없거나 배열이 아닌 경우 예외 처리 수정

* [Fix] 로그인하지 않은 사용자의 경우 조건 수정

* [Style] mt-100 삭제, Footer 사용 시 이전 sibiling에 mb-100 or pb-100 추가 필요

* [Style] 댓글 콘텐츠 영역 whitespace, word-break 설정 수정

* [Fix] 오류 수정

* [Chore] 사용하지 않는 console.* 삭제, 주석 처리

* Merge branch 'dev' into feature/crud

* [Chore] merge error 해결

* [Feat] getStudy() 테스트

* [Fix] /study 접속 시 자동으로 /search? 쿼리 이동 오류 수정. useEffect 제거

+++ src/lib/actions/commentAction.ts
@@ -2,7 +2,8 @@
 
 import { nanoid } from "nanoid";
 import connectDB from "../db";
-import { Comment } from "../schema";
+import { Comment, Post } from "../schema";
+import { revalidateTag } from "next/cache";
 
 // post
 export async function createComment(
@@ -24,16 +25,23 @@ export async function createComment(
   await connectDB();
 
   try {
-    const comment = new Comment({
+    await new Comment({
       postId,
       commentId,
       content,
       writer: userId,
       reply: [],
-    });
+    }).save();
 
-    await comment.save();
+    const newComment = await Comment.findOne({ commentId });
 
+    await Post.findOneAndUpdate(
+      { postId },
+      { $push: { comments: newComment._id } }
+    );
+
+    revalidateTag("community");
+    revalidateTag(postId);
     return { state: true, message: "댓글이 등록 되었습니다." };
   } catch (error) {
     console.log("post comment error" + error);
@@ -42,19 +50,25 @@ export async function createComment(
 }
 
 // get
-export async function getComment(postId: string) {
+export async function getComments(postId: string) {
   await connectDB();
 
   try {
-    const comment = await Comment.findOne({ postId }).populate("writer");
+    const comments = await Comment.find({ postId })
+      .populate("writer", "name role profile_img position_tag")
+      .populate({
+        path: "reply",
+        populate: {
+          path: "writer",
+          select: "name role profile_img position_tag",
+        },
+      })
+      .sort({ createdAt: "asc" });
 
-    return { state: true, data: comment };
+    return { state: true, data: comments };
   } catch (error) {
     console.log("get comment error" + error);
-    return {
-      state: false,
-      message: "댓글 조회에 실패했습니다.",
-    };
+    return { state: false, message: "댓글 조회에 실패했습니다." };
   }
 }
 
@@ -65,13 +79,19 @@ export async function updateComment(commentId: string, formData: FormData) {
   await connectDB();
 
   try {
-    const update = await Comment.findOneAndUpdate({ commentId }, { content });
+    const update = await Comment.findOneAndUpdate(
+      { commentId },
+      { content },
+      { new: true }
+    );
 
     if (!update) {
       return { state: false, message: "해당 댓글을 찾을 수 없습니다." };
     }
 
-    return { state: true, message: "댓글 수정이 완료되었습니다." };
+    revalidateTag("comments");
+    revalidateTag(update.postId);
+    return { state: true, message: "댓글 수정되었습니다." };
   } catch (error) {
     console.log("update comment error" + error);
     return { state: false, message: "댓글 수정에 실패했습니다." };
@@ -83,8 +103,19 @@ export async function deleteComment(commentId: string) {
   await connectDB();
 
   try {
-    await Comment.deleteOne({ commentId });
+    const exist = await Comment.findOne({ commentId });
+    if (exist === undefined || exist === null) {
+      return { success: false, message: "댓글을 찾을 수 없습니다." };
+    }
 
+    await Comment.deleteOne({ commentId });
+    await Post.findOneAndUpdate(
+      { postId: exist.postId },
+      { $pull: { comments: exist._id } }
+    );
+
+    revalidateTag("comments");
+    revalidateTag(exist.postId);
     return { success: true, message: "댓글이 삭제되었습니다." };
   } catch (error) {
     console.error("delete comment error", error);
